移動する床の処理

+++ DirectXGame/Code/game/Collision.cpp
@@ -215,8 +215,8 @@ bool Collision::CollisionRayBox(const XMFLOAT3 startPos, const XMFLOAT3 endPos, 
 
 bool Collision::CollisionBoxPoint(const XMFLOAT3 boxPos, const XMFLOAT3 boxRadius, const XMFLOAT3& pos, const XMFLOAT3 radius)
 {
-	float lengthX = static_cast<float>(fabs(boxPos.x - pos.x));
-	float lengthY = static_cast<float>(fabs(boxPos.y - pos.y));
+	float lengthX = fabsf(boxPos.x - pos.x);
+	float lengthY = fabsf(boxPos.y - pos.y);
 
 	if (lengthX <= boxRadius.x + radius.x && lengthY <= boxRadius.y + radius.y)
 	{
@@ -228,9 +228,9 @@ bool Collision::CollisionBoxPoint(const XMFLOAT3 boxPos, const XMFLOAT3 boxRadiu
 
 bool Collision::CollisionBoxToBox(const XMFLOAT3 boxPos, const XMFLOAT3 boxRadius, const XMFLOAT3& pos, const XMFLOAT3 radius)
 {
-	float lengthX = static_cast<float>(fabs(boxPos.x - pos.x));
-	float lengthY = static_cast<float>(fabs(boxPos.y - pos.y));
-	float lengthZ = static_cast<float>(fabs(boxPos.z - pos.z));
+	float lengthX = fabsf(boxPos.x - pos.x);
+	float lengthY = fabsf(boxPos.y - pos.y);
+	float lengthZ = fabsf(boxPos.z - pos.z);
 
 	if (lengthX < boxRadius.x + radius.x && lengthY < boxRadius.y + radius.y && lengthZ < boxRadius.z + radius.z)
 	{
